build: rename 0c.c file to main.c
In order to make the source code easy to follow by someone who are not
involved on the development, we are renaming the 0c.c file (which is the
compiler main entrypoint) to main.c, so then anyone can guess what is
the role of this file.

Given that we still have a single executable in our code base, I think
it's fine to name the file as main.c.

Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>
Signed-off-by: Carlos Maniero <carlos@maniero.me>