tests: decouple command execution from cli_runner
This commit separates the logic for executing external programs from the
cli_runner. Previously, the abstraction for handling external programs
was tightly coupled with cli_runner.

However, during integration tests, there are numerous instances where we
need to run an external program. To address this, we introduce a new
function, proc_exec, which is dedicated to managing the execution of
external programs.

Signed-off-by: Carlos Maniero <carlos@maniero.me>