cli: add --sysroot and --arch options
In order to compile down into different architectures I am introducing
the --arch and --sysroot options.

The --arch option should enable the user to choose between to
"supported" architectures, which is x86_64 and aarch64.  The default
architecture is x86_64.

The --sysroot option is from where the compiler will find the gas
assembler and the gnu linker.

Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>