cli: add new option to pretty print AST tree
Understanding the intricacies of a parser can be challenging without
adequate visualization tools. This commit addresses this by introducing
a new CLI option to the compiler, enabling the generation of
ASCII-formatted AST trees for easy inspection post-syntax analysis.

It's worth noting that due to current limitations, the tool is optimized
for trees with up to 64 levels.  Beyond 64 levels this tool won't bring
much value I believe.

Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>
Signed-off-by: Carlos Maniero <carlos@maniero.me>