docs: Create olang manual document bootstrap
The idea behind the manual document is to provide a good reference for
anyone who wants to start write projects on olang and contribute to the
language.

We considered to use texinfo since it makes very convenient to
export the documentation as PDF and .info (terminal friendly) files.

This manual document also incorporates the contribution guide previously
documented on olang web site.

Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>