docs: make the olang visuals a w3m like website
This commit makes all font-sizes to be the same, in a way it seems
almost the same in the browser and in the w3m.

I think this is great because we don't need to worrie too much about
font-sizes and enforces the need to build a readable plain text once you
have not much textual style to help you to highlight suff.

Signed-off-by: Carlos Maniero <carlos@maniero.me>
Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>
Message-Id: <20240907033519.293370-4-carlos@maniero.me>