docs: remove HACKING section from README file
The README file is referencing an absent document (HACKING).  Since we
have the full documentation hosted at olang's website and there you can
find "how to build and contribute to the compiler", we can remove the
contribute section from the README and link the documentation.

The documentation can be found at https://o-lang.org and locally for
offline users.  Perhaps in a future change, we can reference the `info`
document on README file to help terminal users to read manual easily.

Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>
Reviewed-by: Carlos Maniero <carlos@maniero.me>
Message-ID: <20240909135320.13021-1-johnny@johnnyrichard.com>