parser: replace type enum to a string view
Ensuring the type is correct is a semantics/checker role. Instead of
trying to convert it to an enum on the parser, it just stores the
identifier of the given type.

Once our spec always expects a TOKEN_COLON (:) before the type
identifier, this constraint was moved to the type parser function.

The string_view is enough until we decide how we gonna handle pointers
and arrays.

Signed-off-by: Carlos Maniero <carlos@maniero.me>
Message-ID: <20240922004619.518310-2-carlos@maniero.me>
Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>