lexer: create token location structure
To reach the source line and column to print errors, we use to get the
filepath from parser, the cursor from token and the source code from the
lexer.

Now, instead of navigating into these three data structures, the same
information must be accessible from a single place, the token's location
*token_loc_t*.

The source_code.h file was removed once the source code struct was added
to the lexer.

Signed-off-by: Carlos Maniero <carlos@maniero.me>
Message-ID: <20241004221157.153364-1-carlos@maniero.me>
Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>