parser: codegen(x86_64): add bitwise not unary op support
Currently, all compiler tests rely on the exit code returned by a binary
program, which restricts our testing to 8-bit unsigned numbers.

To effectively test the bitwise-not unary operation within a main
function, we utilize UINT32_MAX (4294967295), which represents a value
with all bits set. The result of applying the bitwise-not operation to 0
is then used as the exit code.

Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>
Message-ID: <20241009213425.412949-4-johnny@johnnyrichard.com>
Signed-off-by: Carlos Maniero <carlos@maniero.me>