fix: codegen: prevent stack overwrite
There was an issue in the stack allocation algorithm. Consider this
function:

  fn a(): u32 {
    var a: u32 = 0xAAAA
    var b: u64 = 0xBBBBBBBB

    ret
  }

There are three information the stack is required to store:

- 8 bytes: rip (from call instruction)
- 4 bytes: a
- 8 bytes: b

The 0x7FFFFF07 memory address was used to represent the RIP value at
call instant.

Our codegen was assuming the stack works that way:

0       -8  -C
^-------^---^-------
7FFFFF07AAAABBBBBBBB
^-------^---^-------
rip     a   b

So the code gen was:

- Adding the value at the stack;
- Increasing the offset.

But actually the stack was behaving as following:

8       0       -8  -C
^-------^-------^---^
7FFFFF070000BBBBBBBB.
^---------------^---^
rip             a   b

Once the instruction *mov %rax, -0xC(%rbp)* writes from -0xC(%rbp)
(exclusive) to -0x4(%rbp) (inclusive).

So after this change, this is the actual stack template:

        0   -4      -C
--------^---^-------^
7FFFFF07AAAABBBBBBBB.
--------^---^-------^
      rip   a       b

Signed-off-by: Carlos Maniero <carlos@maniero.me>
Message-ID: <20241015121404.206543-1-carlos@maniero.me>
Signed-off-by: Johnny Richard <johnny@johnnyrichard.com>